+++ src/main/java/com/example/elarayax/naves/repository/TallaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.elarayax.naves.model.Talla;
 
 @Repository
-public interface TallaRepository extends JpaRepository<Talla, Integer>{
+public interface TallaRepository extends JpaRepository<Talla, Long>{
 
 }
